Move overlays to the wrapper file
